requirements.txt
python-telegram-bot==20.6
instagrapi==1.19.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
schedule==1.2.1
loguru==0.7.2 